requirements.txt
apipkg==1.4
beautifulsoup4==4.5.3
colorama==0.3.7
coverage==4.0.3
decorator==4.0.10
Django==1.9.13
django-appconf==1.0.1
django-image-cropping==1.0.4
django-recaptcha==1.3.1
django-sendsms==0.2.3
django-webtest==1.7.8
easy-thumbnails==2.3
EasyProcess==0.2.3
execnet==1.4.1
factory-boy==2.8.1
Faker==0.7.10
inflection==0.3.1
ipython==5.0.0
ipython-genutils==0.1.0
isort==4.2.5
pep8==1.7.0
pickleshare==0.7.3
Pillow==3.3.1
prompt-toolkit==1.0.3
psycopg2==2.5.3
py==1.4.33
pygame==1.9.2b1
Pygments==2.1.3
pytest==2.8.3
pytest-cache==1.0
pytest-cov==2.2.0
pytest-django==2.9.1
pytest-env==0.6.0
pytest-factoryboy==1.1.5
pytest-html==1.8.0
pytest-isort==0.1.0
pytest-metadata==1.3.0
pytest-pep8==1.0.6
pytest-sugar==0.5.1
python-dateutil==2.6.0
PyVirtualDisplay==0.2
selenium==2.53.1
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
traitlets==4.2.2
virtualenv==15.1.0
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.2
WebTest==2.0.20
win-unicode-console==0.5
